Backtracking finally complete

+++ pulse/anima/matrix.c
@@ -12,13 +12,13 @@
 // the way we chose our indicies does this automatically.
 // i.e. Inner loop var (in this case k) is not multiplied in said loop.
 void matrix_multiply(float *in1, float *in2, float *out,
-    const unsigned int dim1, const unsigned int dim2, const unsigned int dim3)
+    const int64_t dim1, const int64_t dim2, const int64_t dim3)
 {
-  for (unsigned int i = 0; i < dim1; i++)
+  for (int64_t i = 0; i < dim1; i++)
   {
-    for (unsigned int j = 0; j < dim2; j++)
+    for (int64_t j = 0; j < dim2; j++)
     {
-        for (unsigned int k = 0; k < dim3; k++)
+        for (int64_t k = 0; k < dim3; k++)
         {
             out[i * dim3 + k] += in1[i * dim2 + j] * in2[j * dim3 + k];
         }
